+++ CHANGELOG.md
@@ -5,6 +5,9 @@
     3. [](#bugfix)
         - Fixed "View on GitHub" button in the Changelog modal that was taking you nowhere
         - Equalized blocks sizes are now always rounded to 1 decimal digit and will only be supported this way (fixes #776)
+2. [Joomla](#joomla)
+    3. [](#bugfix)
+        - Object returned by JApplicationSite::getTemplate(true) is not compatible (#499)
 
 # 5.1.2
 ## 31/08/2015